A crude merge of Alexei Toptygin's 20051013 code.
BEHOLD! In the best slack and ass-backwards tradition of rfk
development, the devteam agreed on its mailing list to ship a complete
rewrite of the code done by Alexei Toptygin - in 2005, two years
before Ryan Finnie's 1.4142135.406 release - then everyone (including
Alexei) wandered off and spaced the whole deal.

The key feature of Alexei's rewrite was that NKIs were moved from an
inline string list in the code to an external resource file.

Merging in this rewrite is a bit of a pain because Alexei's file structure
in his tarball didn't match the file structure of the existing code.  This
changeset dumps all of Alexei's C code into robotfindskittens.c and drops
his manual in to replace Ryan Finnie's.  It works, but it's not pretty.

The only functional change from 20051013 in this code is that it looks
for a local "nki" directory after the user's ~/.robotfindskitten directory.
This facilitates testing the program from within the rfk repo without rfk
installed - otherwise it throws a floating-point exception because it can't
find any messages.  This is an undocumented kluge.