Much simpler resize behavior.
Alexey's previous code was a premature optimization - it was working
too hard at avoiding screen redraws, which are fast operations on
modern terminal emulators.  The result was that redraws sometimes
failed to be done when they needed to be.

Mind you, the new resize behavior isn't very interesting - the NKIs don't
move - but at least it's not broken.