Exit 0 upon pressing "q"
Quitting the game is not a failure.  The player is simply choosing not to
play the game.  It's true: robot did not find kitten, so pressing the "q"
key does not result in "success", if that particular goal is how you
define success.  But do we have any right to dictate the conditions of
failure to the player?  Is this a prescriptive zen simulation, or a
descriptive zen simulation?